Fix -Wextra warnings in frameworks/av/include
hardware/qcom/display includes frameworks/av/include headers
while compiling with -Wextra.  Fix warnings that were hidden
by -isystem.

Bug: 31751828
Test: lunch bullhead-userdebug && m -j checkbuild

+++ include/media/Interpolator.h
@@ -152,8 +152,8 @@ public:
                         : mLastSlope;
             } else {
                 // finite difference spline
-                m0 = (sec0 + sec) * 0.5;
-                m1 = (sec1 + sec) * 0.5;
+                m0 = (sec0 + sec) * 0.5f;
+                m1 = (sec1 + sec) * 0.5f;
             }
 
             if (monotonic) {
@@ -294,7 +294,7 @@ public:
 
     std::string toString() const {
         std::stringstream ss;
-        ss << "mInterpolatorType: " << mInterpolatorType << std::endl;
+        ss << "mInterpolatorType: " << static_cast<int32_t>(mInterpolatorType) << std::endl;
         ss << "mFirstSlope: " << mFirstSlope << std::endl;
         ss << "mLastSlope: " << mLastSlope << std::endl;
         for (const auto &pt : *this) {
